requirements.txt
mysql-connector-python==9.1.0
mysqlclient==2.2.6
requests==2.32.3
python-decouple==3.8
python-dotenv==1.0.1
fastapi[standard]==0.115.5
dotmap==1.3.30
pandas==2.2.3
